refactor : 애그리거트 패키지 네이밍 변경
