+++ src/main/java/blackjack/domain/calculator/DefaultScoreStrategy.java
@@ -1,4 +1,4 @@
-package blackjack.domain.score;
+package blackjack.domain.calculator;
 
 import blackjack.domain.card.Card;
